Autonomous and automation
Brought our autonomous back!
- Code has been updated for the new intake

Added failsafe in zero pivot command: if wrist is in a position where the pivot cannot zero against the hardstop and a zero is attempted, the motor will stop if the velocity is close to zero.

Added more automation to scoring (robot remembers if it scored a cone or a cube and acts accordingly.